all test

+++ epam-java-training/webapp/src/main/java/com/chuyeu/training/myapp/webapp/controllers/OrderItemController.java
@@ -15,8 +15,10 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.chuyeu.training.myapp.datamodel.Order;
 import com.chuyeu.training.myapp.datamodel.OrderItem;
 import com.chuyeu.training.myapp.services.IOrderItemService;
+import com.chuyeu.training.myapp.services.IOrderService;
 import com.chuyeu.training.myapp.services.impl.UserAuthStorage;
 import com.chuyeu.training.myapp.webapp.models.OrderItemModel;
 import com.chuyeu.training.myapp.webapp.models.parts.QuantityModel;
@@ -24,62 +26,90 @@ import com.chuyeu.training.myapp.webapp.models.parts.QuantityModel;
 @RestController
 @RequestMapping("/order-item")
 public class OrderItemController {
-	
+
 	@Inject
 	private IOrderItemService orderItemService;
-	
+
 	@Inject
-    private ApplicationContext context;
-	
+	private IOrderService orderService;
+
+	@Inject
+	private ApplicationContext context;
+
 	@RequestMapping(method = RequestMethod.GET)
 	public ResponseEntity<?> getAllByOrderId(@RequestParam(value = "order-id", required = false) Integer orderId) {
 
-		UserAuthStorage userAuthStorage = context.getBean(UserAuthStorage.class);
-		
 		List<OrderItem> listOrderItems = orderItemService.getAllByOrderId(orderId);
 		List<OrderItemModel> listOrderItemsModel = new ArrayList<>();
-		
+
 		for (OrderItem orderItem : listOrderItems) {
 			listOrderItemsModel.add(entyty2model(orderItem));
 		}
 		return new ResponseEntity<List<OrderItemModel>>(listOrderItemsModel, HttpStatus.OK);
 	}
-	
 
 	@RequestMapping(value = "/{id}", method = RequestMethod.GET)
 	public ResponseEntity<?> getById(@PathVariable(value = "id") Integer id) {
 		OrderItem orderItem = orderItemService.get(id);
 		return new ResponseEntity<OrderItemModel>(entyty2model(orderItem), HttpStatus.OK);
 	}
-	
-	
+
 	@RequestMapping(method = RequestMethod.POST)
 	public ResponseEntity<?> createOrderItem(@RequestBody OrderItemModel orderItemModel) {
-		
-		OrderItem orderItem = new OrderItem();
-		orderItem.setProductVariantId(orderItemModel.getProductVariantId());
-		orderItem.setOrderQuantity(orderItemModel.getOrderQuantity());
-		orderItem.setOrderId(orderItemModel.getOrderId());
-		orderItemService.saveOrUpdate(orderItem);
-		
-		return new ResponseEntity<Void>(HttpStatus.CREATED);
+
+		UserAuthStorage userAuthStorage = context.getBean(UserAuthStorage.class);
+		Order order = orderService.get(orderItemModel.getOrderId());
+
+		if (order.getUserProfileId().equals(userAuthStorage.getId())) {
+
+			OrderItem orderItem = new OrderItem();
+			orderItem.setProductVariantId(orderItemModel.getProductVariantId());
+			orderItem.setOrderQuantity(orderItemModel.getOrderQuantity());
+			orderItem.setOrderId(orderItemModel.getOrderId());
+
+			orderItemService.saveOrUpdate(orderItem);
+
+			return new ResponseEntity<Void>(HttpStatus.CREATED);
+		} else {
+			return new ResponseEntity<Void>(HttpStatus.FORBIDDEN);
+		}
 	}
-	
+
 	@RequestMapping(value = "/{id}", method = RequestMethod.PUT)
-	public ResponseEntity<?> updateOrderItem(@RequestBody QuantityModel quantityModel, @PathVariable(value = "id") Integer id) {
+	public ResponseEntity<?> updateOrderItem(@RequestBody QuantityModel quantityModel,
+			@PathVariable(value = "id") Integer id) {
+
 		OrderItem orderItem = orderItemService.get(id);
-		orderItem.setOrderQuantity(quantityModel.getQuantity());
-		orderItemService.saveOrUpdate(orderItem);
-		return new ResponseEntity<Void>(HttpStatus.OK);
+		UserAuthStorage userAuthStorage = context.getBean(UserAuthStorage.class);
+		Order order = orderService.get(orderItem.getOrderId());
+
+		if (order.getUserProfileId().equals(userAuthStorage.getId())) {
+
+			orderItem.setOrderQuantity(quantityModel.getQuantity());
+			orderItemService.saveOrUpdate(orderItem);
+			return new ResponseEntity<Void>(HttpStatus.OK);
+
+		} else {
+			return new ResponseEntity<Void>(HttpStatus.FORBIDDEN);
+		}
 	}
-	
+
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
 	public ResponseEntity<Void> deleteOrderItem(@PathVariable(value = "id") Integer id) {
-		orderItemService.delete(id);
-		return new ResponseEntity<Void>(HttpStatus.OK);
+
+		OrderItem orderItem = orderItemService.get(id);
+		UserAuthStorage userAuthStorage = context.getBean(UserAuthStorage.class);
+		Order order = orderService.get(orderItem.getOrderId());
+
+		if (order.getUserProfileId().equals(userAuthStorage.getId())) {
+			orderItemService.delete(id);
+			return new ResponseEntity<Void>(HttpStatus.OK);
+		} else {
+			return new ResponseEntity<Void>(HttpStatus.FORBIDDEN);
+		}
 	}
-	
-	private OrderItemModel entyty2model(OrderItem orderItem){
+
+	private OrderItemModel entyty2model(OrderItem orderItem) {
 		OrderItemModel orderItemModel = new OrderItemModel();
 		orderItemModel.setId(orderItem.getId());
 		orderItemModel.setProductVariantId(orderItem.getProductVariantId());
